Finished porting to Borland C++ 5.5 compiler. Global variable "log" in WinVNC changed to "vnclog". Proper cooperation between threads in "local cursor" code. Correct loading of configuration files where "compresslevel" or "quality" setting were not set. Corrected path to libjpeg library in the vncviewer.dsw. Minor code clean-ups.
git-svn-id: https://tigervnc.svn.sourceforge.net/svnroot/tigervnc@877 3789f03b-4d11-0410-bbf8-ca57d06f2519

+++ orig/trunk/vnc_winsrc/vncviewer/Flasher.cpp
@@ -199,7 +199,7 @@ LRESULT CALLBACK Flasher::WndProc(HWND hwnd, UINT iMsg, WPARAM wParam, LPARAM lP
 					GetClipBox(hrootdc, &rect);
 					int barwidth = (rect.right - rect.left) / 10;
 					int barheight = max( 
-						(rect.bottom - rect.top) / 10,
+						(int)((rect.bottom - rect.top) / 10),
 						FLASHFONTHEIGHT);
 					HFONT oldfont = (HFONT) SelectObject(hrootdc, _this->m_hfont);
